+++ src/lib.rs
@@ -5,4 +5,3 @@ mod errors;
 
 pub use errors::the_error::{TheResult, TheError, TheErrorType};
 pub use errors::system_error_codes::SystemErrorCodes;
-