+++ src/macros/general.rs
@@ -10,9 +10,9 @@ macro_rules! map_to_new_error {
         $crate::TheError {
             error: $crate::TheErrorType::from($error),
             file: file!().to_string(),
-            location: format!("{}:{}", line!(), column!()),
-            datestamp: chrono::Local::now().date_naive().format("%Y-%m-%d").to_string(),
-            timestamp: chrono::Local::now().time().format("%H:%M:%S").to_string(),
+            location: (line!(), column!()),
+            datestamp: chrono::Local::now().date_naive(),
+            timestamp: chrono::Local::now().time()
         }
     }
 }